Tool renamed to AddFlowSNVQuality and model/conf removed
