+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/featuremapping/FlowFeatureMapperArgumentCollection.java
@@ -131,12 +131,4 @@ public class FlowFeatureMapperArgumentCollection implements Serializable{
     @Argument(fullName = "tag-bases-with-adjacent-ref-diff", doc = "In this mode bases that have an adjacent difference from the reference on the same read are not discarded, and tagged with X_ADJACENT_REF_DIFF", optional = true)
     public boolean tagBasesWithAdjacentRefDiff = false;
 
-    /**
-     *  add-apply-snvqr-features - comma separated list of AppleSNVQR feature to add
-     **/
-    @Hidden
-    @Argument(fullName = "add-apply-snvqr-features", doc = "comma separated list of AppleSNVQR feature to add", optional = true)
-    public List<String> addApplySNVQRFeatures = null;
-
-
 }
