+++ src/test/resources/large/apply_snvqr/snv_apply_snvqr_output.sam
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2faf597265476396793b0aa529da9a6247967c04fe049f66cc04177939b7d153
-size 914426
+oid sha256:678d86622f13e687de10cd6e67c379800fa96f45404ff972771151e419c2af2a
+size 6222340
