--- src/test/resources/large/apply_snvqr/snv_apply_snvqr_output_no_model.sam
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f38f323c2825191c56609dc8dc40a669f471cbae8c92608d0ba1b4f30742a256
-size 914385
