--- src/test/resources/large/apply_snvqr/snv_apply_snvqr_output_no_model_no_conf.sam
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:8bf9baec34c9fa9f898c6e738e372cc638c4b26016abbf2770a12f6010f793bf
-size 6222504
